Modificado sistema para obetener el tamaño de las listas (ahora es mucho más eficiente).
Esto no lo subiré al servidor hasta que haga más cambios, pero esta
testeado que funciona bien

+++ src/main/java/es/ficonlan/web/backend/jersey/resources/EventResource.java
@@ -185,7 +185,7 @@ public class EventResource {
 	@Path("/news/published/size/{eventId}/")
 	@GET
 	@Produces(MediaType.APPLICATION_JSON)
-	public int getAllPublishedNewsItemFromEventTam(@HeaderParam("sessionId") String sessionId, @PathParam("eventId") int eventId) throws ServiceException {
+	public long getAllPublishedNewsItemFromEventTam(@HeaderParam("sessionId") String sessionId, @PathParam("eventId") int eventId) throws ServiceException {
 		return eventService.getAllPublishedNewsItemFromEventTam(sessionId,eventId);
 	}
