+++ src/main/java/es/ficonlan/web/backend/model/eventservice/EventService.java
@@ -82,7 +82,7 @@ public interface EventService {
     
     public List<NewsItem>getAllPublishedNewsItemFormEvent(String sessionId, int eventId, int startIndex, int cont) throws ServiceException;
     
-    public int getAllPublishedNewsItemFromEventTam(String sessionId, int eventId) throws ServiceException;
+    public long getAllPublishedNewsItemFromEventTam(String sessionId, int eventId) throws ServiceException;
     
     public List<NewsItem> getLastNews(String sessionId, Calendar dateLimit, boolean onlyOutstandingNews) throws ServiceException;
