+++ src/main/java/es/ficonlan/web/backend/model/eventservice/EventServiceImpl.java
@@ -745,7 +745,7 @@ public class EventServiceImpl implements EventService {
     
     @Override
     @Transactional(readOnly = true)
-    public int getAllPublishedNewsItemFromEventTam(String sessionId, int eventId) throws ServiceException {
+    public long getAllPublishedNewsItemFromEventTam(String sessionId, int eventId) throws ServiceException {
     	if(!SessionManager.exists(sessionId)) throw new ServiceException(ServiceException.INVALID_SESSION);
 		if(!SessionManager.checkPermissions(SessionManager.getSession(sessionId), "getAllPublishedNewsItemFromEventTam")) throw new ServiceException(ServiceException.PERMISSION_DENIED);
     	return newsDao.getAllPublishedNewsItemFromEventTam(eventId);
